Cleaup tests and compiler warnings
