+++ Cargo.toml
@@ -16,7 +16,7 @@ path = "src/lib.rs"
 [features]
 default = ["bn254"]
 all = [
-  "anssi","bls24","bls48","bls381","bls383","bls461","bn254","bn254CX",
+  "anssi","bls24","bls48","bls381","bls383","bls461","bn254","bn254cx",
   "brainpool","c25519","c41417","ed25519","fp256BN","fp512BN","goldilocks","hifive",
   "nist256","nist384","nist521","nums256e","nums256w","nums384e","nums384w","nums512e",
   "nums512w","rsa2048","rsa3072","rsa4096","secp256k1",
@@ -28,7 +28,7 @@ bls381 = []
 bls383 = []
 bls461 = []
 bn254 = []
-bn254CX = []
+bn254cx = []
 brainpool = []
 c25519 = []
 c41417 = []
