+++ src/ecdh.rs
@@ -745,3 +745,159 @@ pub fn ecies_decrypt(
 
     m
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::test_utils::*;
+    use crate::types::CurveType;
+
+    #[test]
+    fn test_ecdh() {
+
+        let mut rng = create_rng();
+
+        let pw = "M0ng00se";
+        let pp: &[u8] = b"M0ng00se";
+        const EAS: usize = ecp::AESKEY;
+
+        let sha = ecp::HASH_TYPE;
+        let mut salt: [u8; 8] = [0; 8];
+        let mut s1: [u8; EGS] = [0; EGS];
+        let mut w0: [u8; 2 * EFS + 1] = [0; 2 * EFS + 1];
+        let mut w1: [u8; 2 * EFS + 1] = [0; 2 * EFS + 1];
+        let mut z0: [u8; EFS] = [0; EFS];
+        let mut z1: [u8; EFS] = [0; EFS];
+        let mut key: [u8; EAS] = [0; EAS];
+        let mut cs: [u8; EGS] = [0; EGS];
+        let mut ds: [u8; EGS] = [0; EGS];
+        let mut m: Vec<u8> = vec![0; 32]; // array that could be of any length. So use heap.
+        let mut p1: [u8; 3] = [0; 3];
+        let mut p2: [u8; 4] = [0; 4];
+        let mut v: [u8; 2 * EFS + 1] = [0; 2 * EFS + 1];
+        let mut t: [u8; 12] = [0; 12];
+
+        for i in 0..8 {
+            salt[i] = (i + 1) as u8
+        } // set Salt
+
+        println!("\nTesting ECDH/ECDSA/ECIES");
+        println!("Alice's Passphrase= {}", pw);
+
+        let mut s0: [u8; EFS] = [0; EGS];
+        pbkdf2(sha, pp, &salt, 1000, EGS, &mut s0);
+
+        print!("Alice's private key= 0x");
+        printbinary(&s0);
+
+        /* Generate Key pair S/W */
+        key_pair_generate(None, &mut s0, &mut w0);
+
+        print!("Alice's public key= 0x");
+        printbinary(&w0);
+
+        let mut res = public_key_validate(&w0);
+        if res != 0 {
+            println!("ECP Public Key is invalid!");
+            return;
+        }
+
+        /* Random private key for other party */
+        key_pair_generate(Some(&mut rng), &mut s1, &mut w1);
+
+        print!("Servers private key= 0x");
+        printbinary(&s1);
+
+        print!("Servers public key= 0x");
+        printbinary(&w1);
+
+        res = public_key_validate(&w1);
+        if res != 0 {
+            println!("ECP Public Key is invalid!");
+            return;
+        }
+        /* Calculate common key using DH - IEEE 1363 method */
+
+        ecpsvdp_dh(&s0, &w1, &mut z0);
+        ecpsvdp_dh(&s1, &w0, &mut z1);
+
+        let mut same = true;
+        for i in 0..EFS {
+            if z0[i] != z1[i] {
+                same = false
+            }
+        }
+
+        if !same {
+            println!("*** ECPSVDP-DH Failed");
+            return;
+        }
+
+        kdf2(sha, &z0, None, EAS, &mut key);
+
+        print!("Alice's DH Key=  0x");
+        printbinary(&key);
+        print!("Servers DH Key=  0x");
+        printbinary(&key);
+
+        if ecp::CURVETYPE != CurveType::Montgomery {
+            for i in 0..17 {
+                m[i] = i as u8
+            }
+
+            println!("Testing ECIES");
+
+            p1[0] = 0x0;
+            p1[1] = 0x1;
+            p1[2] = 0x2;
+            p2[0] = 0x0;
+            p2[1] = 0x1;
+            p2[2] = 0x2;
+            p2[3] = 0x3;
+
+            let cc = ecies_encrypt(sha, &p1, &p2, &mut rng, &w1, &m[0..17], &mut v, &mut t);
+
+            if let Some(mut c) = cc {
+                println!("Ciphertext= ");
+                print!("V= 0x");
+                printbinary(&v);
+                print!("C= 0x");
+                printbinary(&c);
+                print!("T= 0x");
+                printbinary(&t);
+
+                let mm = ecies_decrypt(sha, &p1, &p2, &v, &mut c, &t, &s1);
+                if let Some(rm) = mm {
+                    println!("Decryption succeeded");
+                    print!("Message is 0x");
+                    printbinary(&rm);
+                } else {
+                    println!("*** ECIES Decryption Failed");
+                    return;
+                }
+            } else {
+                println!("*** ECIES Encryption Failed");
+                return;
+            }
+
+            println!("Testing ECDSA");
+
+            if ecpsp_dsa(sha, &mut rng, &s0, &m[0..17], &mut cs, &mut ds) != 0 {
+                println!("***ECDSA Signature Failed");
+                return;
+            }
+            println!("Signature= ");
+            print!("C= 0x");
+            printbinary(&cs);
+            print!("D= 0x");
+            printbinary(&ds);
+
+            if ecpvp_dsa(sha, &w0, &m[0..17], &cs, &ds) != 0 {
+                println!("***ECDSA Verification Failed");
+                return;
+            } else {
+                println!("ECDSA Signature/Verification succeeded ")
+            }
+        }
+    }
+}
