+++ src/mpin.rs
@@ -351,7 +351,7 @@ pub fn recombine_g2(w1: &[u8], w2: &[u8], w: &mut [u8]) -> isize {
 /* create random secret S */
 pub fn random_generate(rng: &mut RAND, s: &mut [u8]) -> isize {
     let r = Big::new_ints(&rom::CURVE_ORDER);
-    let mut sc = Big::randomnum(&r, rng);
+    let sc = Big::randomnum(&r, rng);
     sc.tobytes(s);
     return 0;
 }
@@ -941,4 +941,228 @@ pub fn server_key(
     hash(sha, &mut c, &mut U, sk);
 
     return 0;
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::test_utils::*;
+    use std::io;
+
+    #[test]
+    fn test_mpin() {
+        let mut rng = create_rng();
+
+        pub const PERMITS: bool = true;
+        pub const PINERROR: bool = true;
+        pub const FULL: bool = true;
+
+        let mut s: [u8; EGS] = [0; EGS];
+        const RM: usize = EFS as usize;
+        let mut hcid: [u8; RM] = [0; RM];
+        let mut hsid: [u8; RM] = [0; RM];
+
+        const G1S: usize = 2 * EFS + 1; /* Group 1 Size */
+        const G2S: usize = 4 * EFS; /* Group 2 Size */
+        const EAS: usize = ecp::AESKEY;
+
+        let mut sst: [u8; G2S] = [0; G2S];
+        let mut token: [u8; G1S] = [0; G1S];
+        let mut permit: [u8; G1S] = [0; G1S];
+        let mut g1: [u8; 12 * EFS] = [0; 12 * EFS];
+        let mut g2: [u8; 12 * EFS] = [0; 12 * EFS];
+        let mut xid: [u8; G1S] = [0; G1S];
+        let mut xcid: [u8; G1S] = [0; G1S];
+        let mut x: [u8; EGS] = [0; EGS];
+        let mut y: [u8; EGS] = [0; EGS];
+        let mut sec: [u8; G1S] = [0; G1S];
+        let mut r: [u8; EGS] = [0; EGS];
+        let mut z: [u8; G1S] = [0; G1S];
+        let mut hid: [u8; G1S] = [0; G1S];
+        let mut htid: [u8; G1S] = [0; G1S];
+        let mut rhid: [u8; G1S] = [0; G1S];
+        let mut w: [u8; EGS] = [0; EGS];
+        let mut t: [u8; G1S] = [0; G1S];
+        let mut e: [u8; 12 * EFS] = [0; 12 * EFS];
+        let mut f: [u8; 12 * EFS] = [0; 12 * EFS];
+        let mut h: [u8; RM] = [0; RM];
+        let mut ck: [u8; EAS] = [0; EAS];
+        let mut sk: [u8; EAS] = [0; EAS];
+
+        let sha = ecp::HASH_TYPE;
+
+        println!("\nTesting MPIN - PIN is 1234");
+        /* Trusted Authority set-up */
+
+        random_generate(&mut rng, &mut s);
+        print!("Master Secret s: 0x");
+        printbinary(&s);
+
+        /* Create Client Identity */
+        let name = "testUser@miracl.com";
+        let client_id = name.as_bytes();
+
+        print!("Client ID= ");
+        printbinary(&client_id);
+
+        hash_id(sha, &client_id, &mut hcid); /* Either Client or TA calculates Hash(ID) - you decide! */
+
+        /* Client and Server are issued secrets by DTA */
+        get_server_secret(&s, &mut sst);
+        print!("Server Secret SS: 0x");
+        printbinary(&sst);
+
+        get_client_secret(&mut s, &hcid, &mut token);
+        print!("Client Secret CS: 0x");
+        printbinary(&token);
+
+        /* Client extracts PIN from secret to create Token */
+        let pin: i32 = 1234;
+        println!("Client extracts PIN= {}", pin);
+        let mut rtn = extract_pin(sha, &client_id, pin, &mut token);
+        if rtn != 0 {
+            println!("FAILURE: EXTRACT_PIN rtn: {}", rtn);
+        }
+
+        print!("Client Token TK: 0x");
+        printbinary(&token);
+
+        if FULL {
+            precompute(&token, &hcid, &mut g1, &mut g2);
+        }
+
+        let mut date = 0;
+        if PERMITS {
+            date = today();
+            /* Client gets "Time Token" permit from DTA */
+
+            get_client_permit(sha, date, &s, &hcid, &mut permit);
+            print!("Time Permit TP: 0x");
+            printbinary(&permit);
+
+            /* This encoding makes Time permit look random - Elligator squared */
+            encoding(&mut rng, &mut permit);
+            print!("Encoded Time Permit TP: 0x");
+            printbinary(&permit);
+            decoding(&mut permit);
+            print!("Decoded Time Permit TP: 0x");
+            printbinary(&permit);
+        }
+
+        print!("\nPIN= ");
+        let _ = io::Write::flush(&mut io::stdout());
+        let mut input_text = String::new();
+        let _ = io::stdin().read_line(&mut input_text);
+
+        let pin = input_text.trim().parse::<usize>().unwrap();
+
+        println!("MPIN Multi Pass");
+        /* Send U=x.ID to server, and recreate secret from token and pin */
+        rtn = client_1(
+            sha,
+            date,
+            &client_id,
+            Some(&mut rng),
+            &mut x,
+            pin,
+            &token,
+            &mut sec,
+            Some(&mut xid[..]),
+            Some(&mut xcid[..]),
+            Some(&permit[..]),
+        );
+        if rtn != 0 {
+            println!("FAILURE: CLIENT_1 rtn: {}", rtn);
+        }
+
+        if FULL {
+            hash_id(sha, &client_id, &mut hcid);
+            get_g1_multiple(Some(&mut rng), 1, &mut r, &hcid, &mut z); /* Also Send Z=r.ID to Server, remember random r */
+        }
+
+        /* Server calculates H(ID) and H(T|H(ID)) (if time PERMITS enabled), and maps them to points on the curve HID and HTID resp. */
+
+        server_1(sha, date, &client_id, &mut hid, Some(&mut htid[..]));
+
+        if date != 0 {
+            rhid.clone_from_slice(&htid[..]);
+        } else {
+            rhid.clone_from_slice(&hid[..]);
+        }
+
+        /* Server generates Random number Y and sends it to Client */
+        random_generate(&mut rng, &mut y);
+
+        if FULL {
+            hash_id(sha, &client_id, &mut hsid);
+            get_g1_multiple(Some(&mut rng), 0, &mut w, &rhid, &mut t); /* Also send T=w.ID to client, remember random w  */
+        }
+
+        /* Client Second Pass: Inputs Client secret SEC, x and y. Outputs -(x+y)*SEC */
+        rtn = client_2(&x, &y, &mut sec);
+        if rtn != 0 {
+            println!("FAILURE: CLIENT_2 rtn: {}", rtn);
+        }
+
+        /* Server Second pass. Inputs hashed client id, random Y, -(x+y)*SEC, xID and xCID and Server secret SST. E and F help kangaroos to find error. */
+        /* If PIN error not required, set E and F = null */
+
+        if !PINERROR {
+            rtn = server_2(
+                date,
+                &hid,
+                Some(&htid[..]),
+                &y,
+                &sst,
+                Some(&xid[..]),
+                Some(&xcid[..]),
+                &sec,
+                None,
+                None,
+            );
+        } else {
+            rtn = server_2(
+                date,
+                &hid,
+                Some(&htid[..]),
+                &y,
+                &sst,
+                Some(&xid[..]),
+                Some(&xcid[..]),
+                &sec,
+                Some(&mut e),
+                Some(&mut f),
+            );
+        }
+
+        if rtn == BAD_PIN {
+            println!("Server says - Bad Pin. I don't know you. Feck off.");
+            if PINERROR {
+                let err = kangaroo(&e, &f);
+                if err != 0 {
+                    println!("(Client PIN is out by {})", err)
+                }
+            }
+            return;
+        } else {
+            println!("Server says - PIN is good! You really are {}", name);
+        }
+
+        if FULL {
+            let mut pxcid = None;
+            if PERMITS {
+                pxcid = Some(&xcid[..])
+            };
+
+            hash_all(sha, &hcid, &xid, pxcid, &sec, &y, &z, &t, &mut h);
+            client_key(sha, &g1, &g2, pin, &r, &x, &h, &t, &mut ck);
+            print!("Client Key =  0x");
+            printbinary(&ck);
+
+            hash_all(sha, &hsid, &xid, pxcid, &sec, &y, &z, &t, &mut h);
+            server_key(sha, &z, &sst, &w, &h, &hid, &xid, pxcid, &mut sk);
+            print!("Server Key =  0x");
+            printbinary(&sk);
+        }
+    }
 }
