+++ src/roms/rom_bn254cx_64.rs
@@ -18,11 +18,11 @@ under the License.
 */
 
 use super::super::arch::Chunk;
-use bn254CX::big::NLEN;
+use bn254cx::big::NLEN;
 use types::{CurvePairingType, CurveType, ModType, SexticTwist, SignOfX};
 
 // Base Bits= 56
-// bn254CX Modulus
+// bn254cx Modulus
 pub const MODULUS: [Chunk; NLEN] = [
     0x6623EF5C1B55B3,
     0xD6EE18093EE1BE,
@@ -53,7 +53,7 @@ pub const FRB: [Chunk; NLEN] = [
     0x10A6F7D0,
 ];
 
-// bn254CX Curve
+// bn254cx Curve
 pub const CURVE_COF_I: isize = 1;
 pub const CURVE_A: isize = 0;
 pub const CURVE_B_I: isize = 2;
